requirements.txt
streamlit==1.32.0
numpy==1.26.4
pretty_midi==0.2.10
python-dotenv==1.0.1
pydantic==2.6.1
mido==1.3.0
python-rtmidi==1.5.8 